[Project: Computational Phantoms] Initial commit
This commit marks the inception of the "Computational Phantoms" repository,
dedicated to the development of algorithms for generating computational
phantom images.

- **README.md**: This file serves as the project's main documentation. It
  provides a comprehensive overview of the project's objectives, the methodology
employed in image generation, and an outline of the files and their roles within
the project. It also includes essential information on installation, usage
instructions, and any prerequisites for utilizing the computational phantom
image generation system.

- **Make_Phantom.m**: This function is the heart of the computational phantom
  image generation process. It's responsible for creating a sequence of DICOM
images that represent the computational phantom.

- **MRBRAIN.DCM**: This DICOM file serves as the template for all MRI DICOM
  images generated by the Make_Phantom.m function. It encapsulates the necessary
metadata and structure required for a valid DICOM image. The image pixel data is
replaced with computed values based on the phantom's properties during the image
generation process.

Signed-off-by: Ljubomir Kurij <ljubomir_kurij@protonmail.com>